Move context strings to constants file
This makes it easier to see which of these exist, and to change them as
versions change. Automatically changing them based on versions isn't a
smart idea, since you may want to keep compatability of proofs between
different versions. On the other hand, a major version might change
things, so this still puts that in the string.